Seamonkey seems to prefer USER_SHEET to AUTHOR_SHEET
The Mozilla documentation also speaks about USER_SHEET, but not
AUTHOR_SHEET.

That, at least, fixes the main functionnality of Stylem, which
is : applying styles to a page !
This makes the extension useable again with Seamonkey.

There's still some bugs like not being able to see all the styles
that apply to a domain.

Still, styles are being applied again and persist through reboots.

Good !

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>